+++ src/FactorGraphTypes.jl
@@ -37,14 +37,14 @@ mutable struct FactorGraph
   bn
   IDs::Dict{Symbol,Int}
   fIDs::Dict{Symbol,Int}
-  id::Int64
+  id::Int
   nodeIDs::Array{Int,1} # TODO -- ordering seems improved to use adj permutation -- pending merge JuliaArchive/Graphs.jl/#225
   factorIDs::Array{Int,1}
   bnverts::Dict{Int,Graphs.ExVertex} # TODO -- not sure if this is still used, remove
   bnid::Int # TODO -- not sure if this is still used
-  dimID::Int64
+  dimID::Int
   cg
-  cgIDs::Dict{Int64,Int64} # cgIDs[exvid] = neoid
+  cgIDs::Dict{Int,Int} # cgIDs[exvid] = neoid
   sessionname::AbstractString
   registeredModuleFunctions::VoidUnion{Dict{Symbol, Function}}
   reference::VoidUnion{Dict{Symbol, Tuple{Symbol, Vector{Float64}}}}
@@ -82,7 +82,7 @@ function emptyFactorGraph(;reference::VoidUnion{Dict{Symbol, Tuple{Symbol, Vecto
                      0,
                      0,
                      nothing,
-                     Dict{Int64,Int64}(),
+                     Dict{Int,Int}(),
                      "",
                      Dict{Symbol, Function}(:IncrementalInference=>IncrementalInference.getSample), # TODO likely to be removed
                      reference  ) #evalPotential
@@ -94,12 +94,12 @@ mutable struct VariableNodeData
   initstdev::Array{Float64,2}
   val::Array{Float64,2}
   bw::Array{Float64,2}
-  BayesNetOutVertIDs::Array{Int64,1}
-  dimIDs::Array{Int64,1}
-  dims::Int64
+  BayesNetOutVertIDs::Array{Int,1}
+  dimIDs::Array{Int,1}
+  dims::Int
   eliminated::Bool
-  BayesNetVertID::Int64
-  separator::Array{Int64,1}
+  BayesNetVertID::Int
+  separator::Array{Int,1}
   groundtruth::VoidUnion{ Dict{ Tuple{Symbol, Vector{Float64}} } }
   VariableNodeData() = new()
   VariableNodeData(x...) = new(x[1],x[2],x[3],x[4],x[5],x[6],x[7],x[8],x[9],x[10],x[11])
@@ -107,19 +107,19 @@ end
 
 mutable struct PackedVariableNodeData
   vecinitval::Array{Float64,1}
-  diminitval::Int64
+  diminitval::Int
   vecinitstdev::Array{Float64,1}
-  diminitdev::Int64
+  diminitdev::Int
   vecval::Array{Float64,1}
-  dimval::Int64
+  dimval::Int
   vecbw::Array{Float64,1}
-  dimbw::Int64
-  BayesNetOutVertIDs::Array{Int64,1}
-  dimIDs::Array{Int64,1}
-  dims::Int64
+  dimbw::Int
+  BayesNetOutVertIDs::Array{Int,1}
+  dimIDs::Array{Int,1}
+  dims::Int
   eliminated::Bool
-  BayesNetVertID::Int64
-  separator::Array{Int64,1}
+  BayesNetVertID::Int
+  separator::Array{Int,1}
   PackedVariableNodeData() = new()
   PackedVariableNodeData(x...) = new(x[1],x[2],x[3],x[4],x[5],x[6],x[7],x[8],x[9],x[10],x[11],x[12],x[13],x[14])
 end
@@ -154,10 +154,10 @@ mutable struct FastRootGenericWrapParam{T} <: Function
 end
 
 mutable struct GenericFunctionNodeData{T, S}
-  fncargvID::Array{Int64,1}
+  fncargvID::Array{Int,1}
   eliminated::Bool
   potentialused::Bool
-  edgeIDs::Array{Int64,1}
+  edgeIDs::Array{Int,1}
   frommodule::S #Union{Symbol, AbstractString}
   fnc::T
   GenericFunctionNodeData{T, S}() where {T, S} = new()
@@ -291,7 +291,7 @@ function addGraphsVert!(fgl::FactorGraph,
   Graphs.add_vertex!(fgl.g, exvert)
 end
 
-function getVertNode(fgl::FactorGraph, id::Int64; nt::Symbol=:var, bigData::Bool=false)
+function getVertNode(fgl::FactorGraph, id::Int; nt::Symbol=:var, bigData::Bool=false)
   return fgl.g.vertices[id] # check equivalence between fgl.v/f[i] and fgl.g.vertices[i]
   # return nt == :var ? fgl.v[id] : fgl.f[id]
 end
@@ -324,11 +324,11 @@ end
 function graphsOutNeighbors(fgl::FactorGraph, vert::Graphs.ExVertex; ready::Int=1,backendset::Int=1, needdata::Bool=false)
   Graphs.out_neighbors(vert, fgl.g)
 end
-function graphsOutNeighbors(fgl::FactorGraph, exVertId::Int64; ready::Int=1,backendset::Int=1, needdata::Bool=false)
+function graphsOutNeighbors(fgl::FactorGraph, exVertId::Int; ready::Int=1,backendset::Int=1, needdata::Bool=false)
   graphsOutNeighbors(fgl.g, getVert(fgl,exVertId), ready=ready, backendset=backendset, needdata=needdata)
 end
 
-function graphsGetEdge(fgl::FactorGraph, id::Int64)
+function graphsGetEdge(fgl::FactorGraph, id::Int)
   nothing
 end
